move deep dive countries tab to individual module
